DeepTrader/requirements.txt
torch>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
pyyaml>=6.0
onnx>=1.14.0
onnxruntime>=1.15.0
tensorboard>=2.14.0
